feat(sue): add reverseGeocode as a separate hook
- moved the `reverseGeocode` hook to the `hooks` folder for easy use elsewhere
- updated in `position` `ImageSelector` instead of `lastKnowPosition` as priority

SUE-76
